+++ tests/units/feature_selection/test_significance_tests.py
@@ -1,3 +1,6 @@
+# -*- coding: utf-8 -*-
+# This file as well as the whole tsfresh package are licenced under the MIT licence (see the LICENCE.txt)
+# Maximilian Christ (maximilianchrist.com), Blue Yonder Gmbh, 2016
 import pytest
 import numpy as np
 import pandas as pd